Login and Registration API
Closes #5, #6

Implemeneted the login and registration API in the python backend. This
addes the `/api/login` and `/api/register` routes that consume a POST
request with the data. They return a json blob that conveys any error or
success status to the client.

Error handling still needs to done and the API's need to be documented
somewhere.